# specs/001-anchor-durable-execution-runtime/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Anchor — durable execution runtime for AI agents
  version: 1.0.0
  description: >-
    Target HTTP contract. Authority is anchor-spec.md §8 for the surface and §31 for what each
    deployment mode exposes. Mode is determined at process start by configuration, never by a
    request, a session, or a user — there is no authentication anywhere in this API, and no
    endpoint accepts or returns a credential.

    Two rules govern the whole document. First, the agent-registration route is NOT MOUNTED in
    demonstration mode and returns 404 rather than 401 or 403, because a response implying that a
    credential would help is a lie about the security model (§27.3). Second, every write endpoint
    that could affect another visitor's data is either absent or scoped to demo runs.
servers:
  - url: http://localhost:8000
    description: Local mode (docker compose; ANCHOR_AUTHORING_EXECUTE=true)
  - url: https://anchor.example.com
    description: Demonstration mode (public; variable unset)
...
paths:
  /api/runs:
    post:
      tags: [runs]
      summary: Submit a run
      description: >-
        Deduplicated on client_request_key. Rate-limited by IP and capped hourly in demonstration
        mode. Appends RUN_SUBMITTED so that submission itself is recoverable from the log.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RunSubmission' }
      responses:
        '201':
          description: Run accepted and queued as pending
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '200':
          description: >-
            Existing run returned because client_request_key matched a prior submission. Not an
            error, and deliberately distinguished from 201 so a caller can tell.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '404': { $ref: '#/components/responses/UnknownAgent' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/FailClosed' }
    get:
      tags: [runs]
      summary: List runs
      parameters:
        - name: status
          in: query
          schema:
            type: array
            items: { $ref: '#/components/schemas/RunStatus' }
          style: form
          explode: true
        - name: agent_type
          in: query
          schema: { type: string }
        - name: is_demo
          in: query
          schema: { type: boolean }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: cursor
          in: query
          description: Opaque cursor; keyset pagination on (created_at, id).
          schema: { type: string }
      responses:
        '200':
          description: >-
            Active runs first, then newest first. Each item carries the thread summary the compact
            strand renders from, and the owning-worker column stays populated because the compact
            strand cannot identify which workers touched a run (§24.8).
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/RunListItem' }
                  next_cursor: { type: [string, 'null'] }
        '503': { $ref: '#/components/responses/FailClosed' }
  /api/runs/{run_id}:
    get:
      tags: [runs]
      summary: Run status, current step, and owning worker
      parameters: [{ $ref: '#/components/parameters/RunId' }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/runs/{run_id}/events:
    get:
      tags: [runs]
      summary: Raw event log for one run, paginated
      description: >-
        after_seq is how a WebSocket client backfills exactly what it missed after being dropped
        for exceeding its buffer (§9).
      parameters:
        - { $ref: '#/components/parameters/RunId' }
        - name: after_seq
          in: query
          schema: { type: integer, minimum: 0 }
        - name: type
          in: query
          schema:
            type: array
            items: { $ref: '#/components/schemas/EventType' }
          style: form
          explode: true
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [items, run_id]
                properties:
                  run_id: { type: integer }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/RunEvent' }
                  next_after_seq: { type: [integer, 'null'] }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/runs/{run_id}/effects:
    get:
      tags: [runs]
      summary: demo_effects rows for this run
      description: >-
        The proof surface (§21.5). One row per side-effect execution. A count of 1 beside a timeline
        showing an interruption is a claim a reviewer can verify without trusting the log.
      parameters: [{ $ref: '#/components/parameters/RunId' }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/DemoEffect' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/runs/{run_id}/resolve:
    post:
      tags: [runs]
      summary: Resolve a needs_review run
      description: >-
        Permitted only while status is needs_review, which is leaseless, so this write can never
        race a worker. The write is attributed to worker_id 'operator' at the run's current epoch,
        because a human decision in the uncertainty window belongs in the log, attributably (I8).
      parameters: [{ $ref: '#/components/parameters/RunId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolution]
              properties:
                resolution:
                  type: string
                  enum: [mark_executed, mark_not_executed, retry]
                  description: >-
                    mark_executed writes the missing TOOL_RESULT now, attributed to the operator,
                    and resumes. mark_not_executed authorizes execution on the next resumption
                    without re-consulting the tool's own declared policy. retry re-consults that
                    declared policy from a clean slate (tasks.md T280) — for a tool declared
                    unsafe this re-enters needs_review immediately, since no automatic policy
                    exists for that category. There is no 'guess' option, by design. Finalizing a
                    needs_review run as cancelled ("abandon") is deferred to phase 6, which is
                    where cooperative cancellation (RUN_CANCELLED) is built.
                note: { type: string, description: Recorded on the event. }
      responses:
        '202':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '403':
          description: Non-demo run in demonstration mode.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '409':
          description: Run is not in needs_review.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
components:
  parameters:
    RunId:
      name: run_id
      in: path
      required: true
      schema: { type: integer }
  responses:
    NotFound:
      description: No such resource.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    UnknownAgent:
      description: agent_type is not registered.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ValidationError:
      description: Malformed request.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    RateLimited:
      description: Per-IP rate limit or hourly demo cap exceeded.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    FailClosed:
      description: >-
        Database unreachable. Nothing executes and nothing is queued, because a side effect that
        cannot be recorded must not happen (I7). This is a design choice, not a limitation.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Stable machine code, e.g. unknown_agent, rate_limited, config_assertion_failed.
        message:
          type: string
          description: >-
            States what happened and what to do. No apologies, no vagueness, and never a raw
            exception string.
        detail: { type: object, additionalProperties: true }
    RunStatus:
      type: string
      enum: [pending, running, completed, failed, cancelled, needs_review]
    EventType:
      type: string
      enum:
        - RUN_SUBMITTED
        - RUN_CLAIMED
        - REPLAY_COMPLETED
        - STEP_STARTED
        - LLM_CALLED
        - TOOL_INTENT
        - TOOL_RESULT
        - NONDET_RECORDED
        - STEP_COMPLETED
        - STEP_SKIPPED_ON_REPLAY
        - STEP_FAILED
        - LEASE_RENEWED
        - WORKER_FENCED
        - RUN_COMPLETED
        - RUN_FAILED
        - RUN_CANCELLED
        - RUN_NEEDS_REVIEW
...
    RunSubmission:
      type: object
      required: [agent_type]
      properties:
        agent_type: { type: string }
        input: { type: object, additionalProperties: true, default: {} }
        client_request_key:
          type: string
          description: Submission-deduplication key. A repeat returns 200 with the existing run.
        priority:
          type: integer
          default: 0
          description: Lower is sooner.
        is_demo: { type: boolean, default: false }
    Run:
      type: object
      required: [id, agent_type, status, epoch, created_at, is_demo]
      properties:
        id: { type: integer }
        display_id: { type: string, example: run_47 }
        agent_type: { type: string }
        status: { $ref: '#/components/schemas/RunStatus' }
        epoch: { type: integer }
        owner_worker_id: { type: [string, 'null'] }
        lease_expires_at: { type: [string, 'null'], format: date-time }
        orphaned:
          type: boolean
          description: >-
            Derived, never stored — status is running and the lease has expired. Stored would
            require a writer at the exact moment nobody owns the run.
        current_step_index: { type: [integer, 'null'] }
        step_count: { type: integer }
        attempts: { type: integer }
        priority: { type: integer }
        is_demo: { type: boolean }
        cancel_requested_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
        claimed_at: { type: [string, 'null'], format: date-time }
        finished_at: { type: [string, 'null'], format: date-time }
    RunListItem:
      allOf:
        - $ref: '#/components/schemas/Run'
        - type: object
          required: [summary, segments]
          properties:
            elapsed_ms: { type: integer }
            segments:
              type: array
              description: Minimal per-owner ranges for the compact thread strand in a list row.
              items: { $ref: '#/components/schemas/TimelineSegment' }
            summary: { $ref: '#/components/schemas/RunSummary' }
    RunSummary:
      type: object
      required: [duplicate_side_effects, handoff_count]
      properties:
        duplicate_side_effects:
          type: integer
          description: >-
            The headline number. Renders 0, explicitly and always — never omitted when zero.
        handoff_count: { type: integer }
        recovery_seconds:
          type: [number, 'null']
          description: >-
            Null when handoff_count is 0, and the interface suppresses it entirely rather than
            printing 0.0s. A zero recovery time on a run that never lost a worker is not a
            measurement and invites distrust of the numbers that are real.
        effects_executed: { type: integer }
        replayed_step_count: { type: integer }
    TimelineStep:
      type: object
      required: [step_index, name, status, action_kind, started_at]
      properties:
        step_index: { type: integer }
        name: { type: string }
        status: { type: string, enum: [done, active, pending, failed, skipped_on_replay] }
        action_kind: { type: string, enum: [tool, model, nondet, done] }
        started_at: { type: string, format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
        duration_ms: { type: [integer, 'null'] }
        attempt: { type: integer, default: 1 }
        idempotency_key: { type: [string, 'null'] }
        idempotency_key_display: { type: [string, 'null'], example: 'r47:s1:a91f' }
        executed:
          type: boolean
          description: >-
            False means the step was read back from the log. The console renders this as fill
            weight, not hue, so the distinction survives grayscale and video compression.
    TimelineSegment:
      type: object
      required: [worker_id, started_at, steps]
      properties:
        worker_id: { type: string }
        epoch: { type: integer }
        claim_reason: { type: string, enum: [initial, reclaimed_after_lease_expiry] }
        started_at: { type: string, format: date-time }
        ended_at:
          type: [string, 'null']
          format: date-time
          description: >-
            Null identifies the current owner. This single field drives the kill button's target,
            the active-step styling, and which strand segment is still growing — so it is trusted
            rather than re-derived.
        steps:
          type: array
          items: { $ref: '#/components/schemas/TimelineStep' }
        log:
          type: array
          items:
            type: object
            required: [timestamp, text, level]
            properties:
              timestamp: { type: string }
              text: { type: string }
              level: { type: string, enum: [info, success, warning] }
...
    RunEvent:
      type: object
      required: [run_id, seq, type, payload, epoch, worker_id, created_at]
      properties:
        run_id: { type: integer }
        seq: { type: integer }
        type: { $ref: '#/components/schemas/EventType' }
        payload: { type: object, additionalProperties: true }
        epoch: { type: integer }
        worker_id: { type: string }
        step_index: { type: [integer, 'null'] }
        created_at: { type: string, format: date-time }
    DemoEffect:
      type: object
      required: [id, run_id, step_index, tool_name, idempotency_key, executed_at]
      properties:
        id: { type: integer }
        run_id: { type: integer }
        step_index: { type: integer }
        tool_name: { type: string }
        idempotency_key: { type: string }
        payload: { type: object, additionalProperties: true }
        executed_at: { type: string, format: date-time }
...